DATAREST-216 - Fixed issue where spring method based security was not working with 'findOne()'.
The findOne(Serializable id) method was being called directly instead of the more specific overloaded version.
As it is imposible to cast to the correct type at compile time, reflection must be used in order to call the more specific method. The CrudRepositoryInvoker#invokeFindOne method has therefore been removed, causing the parent ReflectionRepositoryInvoker#invokeFindOne method to be called instead.