DATAREST-117 - @JsonIgnore and other Jackson annotations are ignored
This fix re-works the PersistentEntityJackson2Module serializer and deserializer such that all Jackson annotations are still honored for both serialization and deserialization.
This fix necessarily involves a rewrite of some existing code pertinent to '_link' section deserialization. The rewriten version now implements this link deserialization properly and therby introduces a significant amount of new functionality in this area. It is therefore now possible to link to other entities when performing a PUT or POST.