DATAREST-238: Change JSON deserializer so it ignores _links
_links are a read-only and should basically be ignored during
any inputs. This required removing the deserializer from
PersistentEntityJackson2Module. A side effect is that any entity
would crash based on the missing _links attribute. Instead of
requiring that every entity apply the necessary annotation,
the mapper is configured to not fail on non-existent attributes.
This allowed all the tests to pass while properly handling PUT
operations.