DATAREST-268, DATAREST-267 - Controllers now ignore URI template variables submitted.
When parsing the id value from a URI we now pipe the obtained segment through Spring HATEOAS' UriTemplate to potentially remove optional tempalte variables. These variables might make it into the URI by clients that do not properly expanding URI templates before submitting requests. 

This is effectively a backport of parts of the fix for DATAREST-155 in master.