DATAREST-538 - Expose Jackson converter beans as being TypeConstrained…Converters
Previously, RepositoryRestMvcConfiguration exposed its two
ResourceSupportHttpMessageConverter beans as MappingJackson2HttpMessageConverters.
This made it impossible for Spring Boot support for conditional beans to identify
that they were type-constrained and, therefore, were not a suitable replacement
for a general-purpose Mapping Jackson2HttpMessageConverter bean.

This commit updates RepositoryRestMvcConfiguration to expose both of its Jackson
message converters as TypeConstrinaedMappingJackson2HttpMessageConverter instances.
This will also Spring Boot to identify that their converters are type-constrained
so that it can continue to auto-configure its general purpose converter. This is
important as it allows Boot to provide a converter that honours the user’s
configuration rather than Spring MVC’s default converter which does not. See
spring-projects/spring-boot#2914 for details.