DATAREST-866 - Avoid premature initialization of RepositoryRestMvcConfiguration.
Previously, RepositoryRestMvcConfiguration declared a BeanPostProcessor instance in a non-static bean factory method. That required Spring to instantiate the configuration class and keeping it in uninitialized state, which could in turn cause issues downstream.

We now use a static method so that the container can obtain the bean instance without having to create an instance of the configuration class and thus deferring its initialization.